Workaround issue in iOS 9 & OSX 11 with IPv6 networks.
getaddrinfo() can return a sockaddr with a nil port if it is given a
numeric service name _and_ the device is connected to an IPv6 network.

Workaround is to manually set the port under these circumstances.

See https://developer.apple.com/videos/play/wwdc2015/719/ for how to set
up a Mac for testing IPv6.